requirements.txt
# Core Dependencies
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
pillow==10.0.0
opencv-python==4.8.0.76
scikit-learn==1.3.0
scikit-image==0.21.0
scipy==1.11.2

# Deep Learning
torch==2.0.1
torchvision==0.15.2
tensorflow==2.13.0
keras==2.13.1

# Segmentation Models
segmentation-models-pytorch==0.3.3
albumentations==1.3.1

# Visualization & Metrics
tensorboard==2.13.0
wandb==0.15.8
tqdm==4.66.1

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
jsonschema==4.19.0

# Optional: For Transformers
transformers==4.31.0
timm==0.9.5